Accommodate install_dev_repos in spy-install-subrepo.sh

+++ spy-install-subrepo.sh
@@ -18,7 +18,11 @@ EXTDEPS=$SPYREPO/external-deps
 if [[ -z "$CONDA_DEFAULT_ENV" && -z "$PYENV_VERSION" && -z "$VIRTUAL_ENV" ]]; then
     echo "Do not install subrepos into base environment. Activate an environment first."; exit 1
 fi
-if [[ "$1" = "python-lsp-server" && -e $SPYREPO/pylsp_utils.py ]]; then
-    export SETUPTOOLS_SCM_PRETEND_VERSION=$(python $SPYREPO/pylsp_utils.py)
+if [[ -e "$SPYREPO/install_dev_repos.py" ]]; then
+    python -bb -X dev -W error $SPYREPO/install_dev_repos.py $@
+else
+    if [[ "$1" = "python-lsp-server" && -e $SPYREPO/pylsp_utils.py ]]; then
+        export SETUPTOOLS_SCM_PRETEND_VERSION=$(python $SPYREPO/pylsp_utils.py)
+    fi
+    python -m pip install --no-deps -e $EXTDEPS/$1
 fi
-python -m pip install --no-deps -e $EXTDEPS/$1
